backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-whois==0.8.0
Pillow==10.1.0
reportlab==4.0.7
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
jinja2==3.1.2
httpx==0.25.2
beautifulsoup4==4.12.2
dnspython==2.4.2
shodan==1.31.0
censys==2.2.7
passlib
pyotp
qrcode==2.4.2
yandex-images-downloader==1.0.0
